[feature] Icon Update
Allows for an entire Icon object to be configured from the json file. This is so that we are able to use the current Icon library (Shape.so) that we paid for which relies heavily on Stroke instead of Fill.

All icons that were previously established are still backwards compatible.

Why on Locations? Expand Icon used for the locations cards.