Fixed the FAQ Scroll (#538)
* Fixed the FAQ Scroll

When clicking the see more or see less button it would scroll way above the FAQ Section. Now it scrolls to the right section

* Fix for Mobile

* Revert "Fix for Mobile"

This reverts commit ac58081af93541799d305825fc4035a02fa96d61.

* Added mobile changes

* FIxed the scroll for other locations

I created a function the same way there is one for the data to changed based on campus for the scroll so in the future this can be changed easier for other locations that need adjusting.