Updated Schema and Formatting Season Title
Added custom formatting that formats the Season title from the URL. It works pretty well for the most part but can't grab special characters(like "?"), due to them not being in the URL. Only case we run into is "Season 1 | What's So Good?". We may want to revisit this, but will leave this for time being.